backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
openai>=1.6.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
